requirements.txt
asttokens==2.4.1
attrs==24.3.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
dnspython==2.7.0
et_xmlfile==2.0.0
executing==2.0.1
feedparser==6.0.11
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.55.3
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.66.0
gspread==6.1.4
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jupyter-ui-poll==1.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
msal==1.31.1
nest-asyncio==1.6.0
numpy==2.2.1
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.29.2
psutil==6.0.0
pure_eval==0.2.3
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
pymongo==4.10.1
PyMuPDF==1.25.1
pynput==1.7.7
pyparsing==3.2.0
PyPDF2==3.0.1
PySocks==1.7.1
pyspark==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==306
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
schedule==1.2.2
seaborn==0.13.2
selenium==4.27.1
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
wcwidth==0.2.13
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
wsproto==1.2.0
